Languages.md (#377)
* Update 03. Languages.md

removed this line as it was not needed any more 
For details, you can refer to the article https://help.nyaruka.com/en/articles/4207750-changing-the-default-flow-language published by the Nyaruka team for the community. We thank the Nyaruka team for publishing it publicly for the community.

* Update 03. Languages.md

added languages supported for template messages