Update 03. Send the contact an interactive message.md (#441)
* Update 03. Send the contact an interactive message.md

Updated the entire documentation along with level+read time+ last updated date. 
The section to be clubbed under a different documentation, hence not present here.

* Update 03. Send the contact an interactive message.md

Re-wrote the existing documentation along with adding the last update date, level and read time. 
The dynamic fetching from sheets is to be clubbed with another documentation, hence removed it from this documentation.

* Update 03. Send the contact an interactive message.md

Re-wrote this documentation along with level, last update date and read time. the last part with dynamic fetching from g-sheets is to be clubbed with another existing documentation, hence removed it from here.

* Update 03. Send the contact an interactive message.md

* Update 03. Send the contact an interactive message.md

* Update 03. Send the contact an interactive message.md

* Update 03. Send the contact an interactive message.md

---------

Co-authored-by: Akansha Sakhre <asakhre2002@gmail.com>
Co-authored-by: Sangeet Mishra <143380171+SangeetaMishr@users.noreply.github.com>